Fix Linux build process

+++ build-linux.sh
@@ -11,12 +11,20 @@ set -e
 
 # You will probably need to change the -Fl parameter below to point
 # to where you have indy10
-compiler="fpc -Mdelphi -Tlinux -O3 -vew -Sew -vq -Fu~/sources/indy10/ -Fu/usr/lib64/lazarus/lcl/units/x86_64-linux/ -Fu/usr/lib64/lazarus/components/lazutils/lib/x86_64-linux/"
-
+compiler="fpc -Mdelphi -Tlinux -O3 -vew -vq"
+compiler=$compiler" -Fu/data_b/source/indy10/Lib/*"
+compiler=$compiler" -Fu/usr/share/fpcsrc/2.6.2/packages/fcl-base/src"
+compiler=$compiler" -Fu/usr/share/fpcsrc/2.6.2/packages/fcl-net/src"
+compiler=$compiler" -Fu/usr/share/fpcsrc/2.6.2/packages/iconvenc/src"
+compiler=$compiler" -FE."
+echo $compiler
 
 function compile {
     $compiler $1.dpr
-    mv $1 ../bin
+    if [ -e "$1" ]
+    then
+        mv $1 ../bin
+    fi
     echo -e '\n'
 }
 
@@ -33,6 +41,7 @@ compile Sleep
 compile Parser
 compile GenDb
 compile cgi/query
+mv query ../bin
 #compile Robot
 #compile SearchServer
